pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "diamondhacks"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "numpy",
    "pandas",
    "scikit-learn",
    "rdkit-pypi>=2022.9.1",
    "selfies>=2.1",
    "pyyaml",
    "omegaconf",
    "tqdm",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "requests",
]

[project.optional-dependencies]
dev = ["pytest", "wandb"]
mist = ["mist @ git+https://github.com/samgoldman97/mist.git@master"]

[tool.setuptools.packages.find]
include = ["src*", "backend*"]
